+++ tests/cli/commands/test_add_raster.py
@@ -9,7 +9,7 @@ from tests.conftest import expected_json
 def test_add_raster_to_items(tmp_planet_disaster: pystac.Collection):
     collection = tmp_planet_disaster
     collection_path = collection.get_self_href()
-    items = list(collection.get_all_items())
+    items = list(collection.get_items(recursive=True))
     item_path = pystac.utils.make_absolute_href(
         items[0].get_self_href(), collection_path
     )
@@ -19,7 +19,7 @@ def test_add_raster_to_items(tmp_planet_disaster: pystac.Collection):
     assert result.exit_code == 0
 
     updated = pystac.read_file(collection_path)
-    item = list(updated.get_all_items())[0]
+    item = list(updated.get_items(recursive=True))[0]
     asset = item.get_assets().get("analytic")
     assert asset is not None
     expected = expected_json("rasterbands.json")